Create  Security.md
a.Created Security.md
b. Describe your changes.
Created Security.md to ensure the security and integrity of the ezmail repository, it's essential to establish a comprehensive security policy. This policy will outline guidelines, best practices, and measures to safeguard the codebase, user data, and the project as a whole.
c. Put closes #XXXX in your comment to auto-close the issue that your PR fixes (if such).

closes #48
Fixes #48